Added option to block accounts
Added the option to block/unblock another user, the feature is a button next to the follow/unfollow button in 'profiles.show'.
When you block or get blocked by another user, you can no longer:
• follow them
• see their articles
• see their comments
• send them new messages
This is done using another pivot table on 'profile_user' with a different pivot table name to differentiate it from the 'follow' pivot table on the same two models.

Additional notes:
• Later, I will update the restrictions implemented by the blocking feature as Laravel Policies, makes more sense and would render operations more secure.