Added tags to articles
Now the users have the option to add tags to articles through a pivot table 'article_tag'.
The tags are handled through a Helper function 'handleTags'.
The tags are accessed through a GET route method, similar to categories.

Will probably update the tags as 'React chip components' after learning and integrating React.js to the app